Conditional Kendall's tau

Computes the conditional Kendall's tau and inference

cKendall(trun, obs, delta = NULL, method = "MB", weights = NULL,
  a = 0, trans = "linear", ...)

Arguments

trun
left truncation time satisfying trun <= obs.

obs
observed failure time, must be the same length as trun, might be right-censored.

delta
an optional 0-1 vector of censoring indicator (0 = censored, 1 = event) for obs. If this vector is not specified, cKendall assumes no censoring and all observed failure time denote events.

method
a character string specifying the different version of conditional Kendall's tau to be computed. The following are permitted:

MB

conditional Kendall's tau proposed in Martin and Betensky (2005) as \(\hat{\tau}_c,\)

IPW1

inverse probability weighted estimator proposed in Austin and Betensky (2014) as \(\hat{\tau}_{c2},\)

IPW2

restricted inverse probability weighted estimator proposed in Austin and Betensky (2014) as \(\hat{\tau}_{c3}.\)

weights
an optional vector of sampling weights used when method = IPW1 or method = IPW2. Inverse probability censored weighting (IPCW) is the default.

a
a numeric transformation parameter. The default value is 0, which applies no transformation. This parameter must be greater than -1. See ?tranSurvfit for the transformation model structure.

trans
a character string specifying the transformation structure. The following are permitted:

linear

linear transformation structure,

log

log-linear transformation structure,

exp

exponential transformation structure.

...
for future methods.

Details

This function performs statistical test for quasi-independence between truncation time and failure time. The hypothesis test is based on the conditional Kendall's tau of Martin and Betensky (2005) and the two versions of the inverse probability weighted Kendall's tau of Austin and Betensky (2014).

The output contains the following components:

PE

consistent point estimate of the conditional Kendall's tau.

SE

asymptotic standard error of the conditional Kendall's tau estimator.

STAT

the value of the normal test statistic.

p.value

the (Wald) p-value of the test.

trans

the transformation model (if applied).

a

the estimated transformation parameter.

References

Martin E. and Betensky R. A. (2005), Testing quasi-independence of failure and truncation times via conditional Kendall's tau, Journal of the American Statistical Association, 100 (470): 484-492.

Austin, M. D. and Betensky R. A. (2014), Eliminating bias due to censoring in Kendall's tau estimators for quasi-independence of truncation and failure, Computational Statistics & Data Analysis, 73: 16-26.

See also

Examples

data(channing, package = "boot")
chan <- subset(channing, sex == "Male" & entry < exit)
attach(chan)
cKendall(entry, exit, cens)

#> 
#>  Test for quasi-independence with conditional Kendall's tau
#> 
#>  Call: cKendall(trun = entry, obs = exit, delta = cens)
#> 
#>  Kendall's tau = 0.1967 , SE = 0.0958 , Z = 2.0529 , p-value =  0.0401 
#> 

cKendall(entry, exit, cens, method = "IPW1")

#> 
#>  Test for quasi-independence with conditional Kendall's tau
#> 
#>  Call: cKendall(trun = entry, obs = exit, delta = cens, method = "IPW1")
#> 
#>  Kendall's tau = 0.1356 , SE = 0.0927 , Z = 1.4631 , p-value =  0.1434 
#> 

cKendall(entry, exit, cens, method = "IPW2")

#> 
#>  Test for quasi-independence with conditional Kendall's tau
#> 
#>  Call: cKendall(trun = entry, obs = exit, delta = cens, method = "IPW2")
#> 
#>  Kendall's tau = 0.425 , SE = 0.174 , Z = 2.4423 , p-value =  0.0146 
#> 

detach(chan)